pal/net: convenience headers

+++ pal/net/ip/address.hpp
@@ -128,7 +128,7 @@ public:
 	constexpr auto operator<=> (const address &) const noexcept = default;
 
 	/// Return hash value for \a this
-	[[nodiscard]] constexpr uint64_t hash () const noexcept
+	[[nodiscard]] constexpr size_t hash () const noexcept
 	{
 		if (const auto *p = v4())
 		{
